+++ src/entities/notefolder.cpp
@@ -239,7 +239,8 @@ bool NoteFolder::store() {
     query.bindValue(":showSubfolders", this->showSubfolders);
     query.bindValue(":activeNoteSubFolderData", this->activeNoteSubFolderData);
 
-    // make the path relative to the portable data path if we are in portable mode
+    // make the path relative to the portable data path if we are in
+    // portable mode
     query.bindValue(":localPath",
                     Utils::Misc::makePathRelativeToPortableDataPathIfNeeded(
                             this->localPath));
@@ -272,7 +273,8 @@ void NoteFolder::setAsCurrent() {
     QSettings settings;
     settings.setValue("currentNoteFolderId", id);
 
-    // make the path relative to the portable data path if we are in portable mode
+    // make the path relative to the portable data path if we are in
+    // portable mode
     settings.setValue("notesPath",
                       Utils::Misc::makePathRelativeToPortableDataPathIfNeeded(
                               localPath));